backend/pyproject.toml
[project]
name = "realworld-django"
version = "0.1.0"
requires-python = ">=3.12, <3.13"
dependencies = [
    "django==5.0.4",
    "django-cors-headers==4.3.1",
    "django-filter==24.2",
    "djangorestframework==3.15.1",
    "djangorestframework-simplejwt==5.3.1",
    "drf-nested-routers==0.93.5",
    "mixer==7.2.2",
    "pillow==10.3.0",
]

[dependency-groups]
dev = [
    "ruff==0.3.7",
]
